Implement a hijackable transport type
This wraps a net.Conn and allows it to be used as an http.RoundTripper.
It also removes the need to duplicate code in the authenticator type
(the connect() and do() functions were basically the same), which will
make it easier to add more authentication methods (e.g. Kerberos) later.